Aod_project/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_gis==1.1
earthaccess==0.14.0
numpy==2.2.4
psycopg2_binary==2.9.10
rasterio==1.4.3
Requests==2.32.3
xarray==2025.1.2
django-cors-headers==4.7.0
django-crontab==0.7.1
netCDF4==1.7.2
h5netcdf==1.6.1
rioxarray==0.19.0
bs3==0.0.2
xlrd==2.0.1
joblib==1.5.0
geopandas==1.0.1